Use the same storage folder for all examples
If you’re running through all the examples SDK will be creating a ./.storage_dir on each example folder, so if you go through different examples with the same account it will create different storage files.

This can cause problems when retrieving the messages from the server, as it will report its on a different history position.